+++ src/main/java/net/Core/Physics/CubeCollision.java
@@ -1,18 +1,19 @@
 package net.Core.Physics;
 
+import org.joml.Vector3d;
 import org.joml.Vector3f;
 
 import static org.lwjgl.opengl.GL40.*;
 
 public class CubeCollision {
-    public Vector3f position;
-    public Vector3f min;
-    public Vector3f max;
-    public Vector3f size;
+    public Vector3d position;
+    public Vector3d min;
+    public Vector3d max;
+    public Vector3d size;
 
-    public CubeCollision(Vector3f position, Vector3f size) {
-        this.min = new Vector3f(Math.min(0, size.x), Math.min(0, size.y), Math.min(0, size.z));
-        this.max = new Vector3f(Math.max(0, size.x), Math.max(0, size.y), Math.max(0, size.z));
+    public CubeCollision(Vector3d position, Vector3d size) {
+        this.min = new Vector3d(Math.min(0, size.x), Math.min(0, size.y), Math.min(0, size.z));
+        this.max = new Vector3d(Math.max(0, size.x), Math.max(0, size.y), Math.max(0, size.z));
         this.position = position;
         this.size = size;
     }
@@ -48,24 +49,24 @@ public class CubeCollision {
         glDrawArrays(GL_LINE_LOOP, 0, 36);
     }
 
-    public static Vector3f getIntersectionCenter(CubeCollision a, CubeCollision b) {
-        float interMinX = Math.max(a.min.x, b.min.x);
-        float interMinY = Math.max(a.min.y, b.min.y);
-        float interMinZ = Math.max(a.min.z, b.min.z);
+    public static Vector3d getIntersectionCenter(CubeCollision a, CubeCollision b) {
+        double interMinX = Math.max(a.min.x, b.min.x);
+        double interMinY = Math.max(a.min.y, b.min.y);
+        double interMinZ = Math.max(a.min.z, b.min.z);
 
-        float interMaxX = Math.min(a.max.x, b.max.x);
-        float interMaxY = Math.min(a.max.y, b.max.y);
-        float interMaxZ = Math.min(a.max.z, b.max.z);
+        double interMaxX = Math.min(a.max.x, b.max.x);
+        double interMaxY = Math.min(a.max.y, b.max.y);
+        double interMaxZ = Math.min(a.max.z, b.max.z);
 
-        return new Vector3f(
+        return new Vector3d(
                 (interMinX + interMaxX) / 2.0f,
                 (interMinY + interMaxY) / 2.0f,
                 (interMinZ + interMaxZ) / 2.0f
         );
     }
 
-    public Vector3f getCenter() {
-        return new Vector3f(
+    public Vector3d getCenter() {
+        return new Vector3d(
                 (min.x + max.x) / 2.0f,
                 (min.y + max.y) / 2.0f,
                 (min.z + max.z) / 2.0f
